backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "elevenlabs>=2.1.0",
    "fastapi[standard]>=0.115.12",
    "google-genai>=1.16.1",
    "langchain>=0.3.25",
    "langchain-google-genai>=2.1.4",
    "langgraph>=0.4.7",
    "pillow>=11.2.1",
    "pydantic==2.11.4",
    "supabase>=2.15.1",
]
